updated logging based on code review comments... learning the python ways
